requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
astunparse==1.6.3
async-timeout==4.0.2
attrs==23.1.0
black==23.3.0
cachetools==5.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
contourpy==1.0.7
cycler==0.11.0
flatbuffers==1.12
fonttools==4.39.3
frozenlist==1.3.3
gast==0.4.0
google-auth==2.17.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.54.0
h5py==3.8.0
idna==3.4
imageio==2.28.1
importlib-metadata==6.6.0
importlib-resources==5.12.0
keras==2.9.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
lazy_loader==0.2
libclang==16.0.0
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.1
numpy==1.24.3
oauthlib==3.2.2
opencv-python==4.7.0.72
opt-einsum==3.3.0
packaging==23.1
pathspec==0.11.1
Pillow==9.5.0
platformdirs==3.5.0
protobuf==3.19.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.0.9
python-dateutil==2.8.2
PyWavelets==1.4.1
requests==2.30.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.20.0
scipy==1.9.1
six==1.15.0
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.9.0
tensorflow-macos==2.9.0
tensorflow-metal==0.5.0
termcolor==2.3.0
tifffile==2023.4.12
tomli==2.0.1
typing_extensions==4.5.0
urllib3==2.0.2
Werkzeug==2.3.4
wrapt==1.15.0
yarl==1.9.2
zipp==3.15.0
